WIP
testCreateNativeListener - Removed all the mock infrastructure, cleaned
up some code & fixed ROUTER_PORT_CONNECTION_FAILED TEST

testCreatePeerListener - Moved the tests from testThaliTcpServersManager
that had to do with the createPeerListener and put in a lot of test
stubs for things we need to test but aren't currrently being tested.

testThaliTcpServersManager - Allowed the test to run with coordinator,
some code clean up, put in test for terminateIncomingConnection and
added test stubs we need to fill in

createNativeListener - Fixed ROUTER_PORT_CONNECTION_FAILED syntax

createPeerListener - Moved code from thaliTcpServersManager and made
some minor fixes.

thaliTcpServersManager - Pulled out createPeerListener code and cleaned
up the result. Added code for terminateIncomingConnection and added
stub for terminateOutgoingConnection.
